move audio transcription service to nlp
